Module Protocol.Fees_storage

type Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error +=

| Cannot_pay_storage_fee

type Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error +=

| Operation_quota_exceeded

type Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error +=

| Storage_limit_too_high

val origination_burn : Tezos_raw_protocol_010_PtGRANAD.Raw_context.t -> (Tezos_raw_protocol_010_PtGRANAD.Raw_context.t * Tezos_raw_protocol_010_PtGRANAD.Tez_repr.t, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result

Does not burn, only adds the burn to storage space to be paid

val record_paid_storage_space : Tezos_raw_protocol_010_PtGRANAD.Raw_context.t -> Tezos_raw_protocol_010_PtGRANAD.Contract_repr.t -> (Tezos_raw_protocol_010_PtGRANAD.Raw_context.t * Z.t * Z.t * Tezos_raw_protocol_010_PtGRANAD.Tez_repr.t, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result Lwt.t

The returned Tez quantity is for logging purpose only

val record_paid_storage_space_subsidy : Tezos_raw_protocol_010_PtGRANAD.Raw_context.t -> Tezos_raw_protocol_010_PtGRANAD.Contract_repr.t -> (Tezos_raw_protocol_010_PtGRANAD.Raw_context.t * Z.t * Z.t, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result Lwt.t

Record paid storage space for contract without burn. For use only in subsidies. Will fail if storage_space_to_pay has been initialized.

val check_storage_limit : Tezos_raw_protocol_010_PtGRANAD.Raw_context.t -> storage_limit:Z.t -> (unit, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result

val start_counting_storage_fees : Tezos_raw_protocol_010_PtGRANAD.Raw_context.t -> Tezos_raw_protocol_010_PtGRANAD.Raw_context.t

val burn_storage_fees : Tezos_raw_protocol_010_PtGRANAD.Raw_context.t -> storage_limit:Z.t -> payer:Tezos_raw_protocol_010_PtGRANAD.Contract_repr.t -> (Tezos_raw_protocol_010_PtGRANAD.Raw_context.t, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result Lwt.t