Module Tezos_raw_protocol_010_PtGRANAD.Period_repr

type t

type period = t

A period is a non-negative integer.

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val encoding : period Tezos_base__TzPervasives.Data_encoding.t

val rpc_arg : period Tezos_protocol_environment_010_PtGRANAD__Environment.RPC_arg.t

val pp : Stdlib.Format.formatter -> period -> unit

val to_seconds : period -> int64

val of_seconds : int64 -> (period, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result

of_second period fails if period is not positive

val of_seconds_exn : int64 -> period

of_second period fails if period is not positive. It should only be used at toplevel for constants.

val add : period -> period -> (period, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result

val (+?) : period -> period -> (period, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result

val mult : int32 -> period -> (period, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result

val zero : period

val one_second : period

val one_minute : period

val one_hour : period

val compare : period -> period -> int

compare x y returns 0 if x is equal to y, a negative integer if x is shorter than y, and a positive integer if x is longer than y.