Module Roll_storage.Contract

val add_amount : Raw_context.t -> Contract_repr.t -> Tez_repr.t -> (Raw_context.t, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result Lwt.t

Calls Delegate.add_amount ctxt contract am if a delegate is associated to contract, or returns unchanged ctxt otherwise.

val remove_amount : Raw_context.t -> Contract_repr.t -> Tez_repr.t -> (Raw_context.t, Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.error Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.trace) Stdlib.result Lwt.t

Calls Delegate.remove_amount ctxt contract am if a delegate is associated to contract, or returns unchanged ctxt otherwise.