Module Sapling_storage.Roots

Bounded queue of roots. The full size is initialized with the default uncommitted root, that's why roots storage doesn't need to be carbonated. A maximum of one new root is added per protocol level. If multiple transactions for the same shielded pool are processed during the same contract call or several calls in the same block, only the last root will be stored. This property prevents transactions in the same block from depending on each other and guarantees that a transaction will be valid for a least two hours (hence the 120 size) after being forged.

val size : int32

val get : Raw_context.t -> Tezos_raw_protocol_010_PtGRANAD__Storage.Sapling.id -> (Storage.Sapling.Roots.value, Tezos_protocol_environment_010_PtGRANAD.Error_monad.error Tezos_protocol_environment_010_PtGRANAD.Error_monad.trace) Tezos_protocol_environment_010_PtGRANAD.Pervasives.result Tezos_protocol_environment_010_PtGRANAD.Lwt.t

val init : Raw_context.t -> Tezos_raw_protocol_010_PtGRANAD__Storage.Sapling.id -> (Raw_context.t, Tezos_protocol_environment_010_PtGRANAD.Error_monad.error Tezos_protocol_environment_010_PtGRANAD.Error_monad.trace) Tezos_protocol_environment_010_PtGRANAD.Pervasives.result Tezos_protocol_environment_010_PtGRANAD.Lwt.t

val mem : Raw_context.t -> Tezos_raw_protocol_010_PtGRANAD__Storage.Sapling.id -> Tezos_protocol_environment_010_PtGRANAD.Sapling.Hash.t -> (bool, Tezos_protocol_environment_010_PtGRANAD.Error_monad.error Tezos_protocol_environment_010_PtGRANAD.Error_monad.trace) Tezos_protocol_environment_010_PtGRANAD.Pervasives.result Tezos_protocol_environment_010_PtGRANAD.Lwt.t

val add : Raw_context.t -> Tezos_raw_protocol_010_PtGRANAD__Storage.Sapling.id -> Storage.Sapling.Roots.value -> (Raw_context.t, Tezos_protocol_environment_010_PtGRANAD.Error_monad.error Tezos_protocol_environment_010_PtGRANAD.Error_monad.trace) Tezos_protocol_environment_010_PtGRANAD.Pervasives.result Tezos_protocol_environment_010_PtGRANAD.Lwt.t