Module Storage.Liquidity_baking

module Escape_ema : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Tezos_protocol_environment_010_PtGRANAD__Environment.Int32.t

Exponential moving average (ema) of flags set in protocol_data.contents. If at any block it's above the threshold set in constants, liquidity baking permanently shuts off. *

module Cpmm_address : Storage_sigs.Single_data_storage with type t := Raw_context.t and type value = Contract_repr.t

Constant product market maker contract that receives liquidity baking subsidy. *