Module Tezos_raw_protocol_010_PtGRANAD.Vote_storage

val record_proposal : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Protocol_hash.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.Public_key_hash.t -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

Records a protocol proposal with the delegate that proposed it.

val recorded_proposal_count_for_delegate : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.Public_key_hash.t -> int Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val get_proposals : Raw_context.t -> int32 Tezos_protocol_environment_010_PtGRANAD__Environment.Protocol_hash.Map.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

Computes for each proposal how many delegates proposed it.

val clear_proposals : Raw_context.t -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

type ballots = {

yay : int32;

nay : int32;

pass : int32;

}

Counts of the votes

val ballots_encoding : ballots Tezos_protocol_environment_010_PtGRANAD__Environment.Data_encoding.t

val has_recorded_ballot : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val record_ballot : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.Public_key_hash.t -> Vote_repr.ballot -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

Records a vote for a delegate, returns a Storage_error Existing_key if the vote was already registered

val get_ballots : Raw_context.t -> ballots Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

Computes the sum of the current ballots weighted by stake.

val get_ballot_list : Raw_context.t -> (Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.Public_key_hash.t * Vote_repr.ballot) list Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val clear_ballots : Raw_context.t -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val listings_encoding : (Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.Public_key_hash.t * int32) list Tezos_protocol_environment_010_PtGRANAD__Environment.Data_encoding.t

val update_listings : Raw_context.t -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

Populates !Storage.Vote.Listings using the currently existing rolls and sets Listings_size. Delegates without rolls are not included in the listing.

val listing_size : Raw_context.t -> int32 Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

Returns the sum of all rolls of all delegates.

val in_listings : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

Verifies the presence of a delegate in the listing.

val get_listings : Raw_context.t -> (Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.Public_key_hash.t * int32) list Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val get_voting_power_free : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.public_key_hash -> int32 Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val get_voting_power : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Signature.public_key_hash -> (Raw_context.t * int32) Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val get_total_voting_power_free : Raw_context.t -> int32 Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val get_total_voting_power : Raw_context.t -> (Raw_context.t * int32) Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val get_current_quorum : Raw_context.t -> int32 Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val get_participation_ema : Raw_context.t -> int32 Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val set_participation_ema : Raw_context.t -> int32 -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val get_current_proposal : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Protocol_hash.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val find_current_proposal : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Protocol_hash.t option Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val init_current_proposal : Raw_context.t -> Tezos_protocol_environment_010_PtGRANAD__Environment.Protocol_hash.t -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val clear_current_proposal : Raw_context.t -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

val init : Raw_context.t -> start_position:Tezos_protocol_environment_010_PtGRANAD__Environment.Int32.t -> Raw_context.t Tezos_protocol_environment_010_PtGRANAD__Environment.Error_monad.tzresult Tezos_protocol_environment_010_PtGRANAD__Environment.Lwt.t

Sets the initial quorum to 80% and period kind to proposal.