Module Tezos_raw_protocol_010_PtGRANAD__Amendment

Only delegates with at least one roll take part in the amendment procedure. It works as follows:

Proposal period: delegates can submit protocol amendment proposals using the proposal operation. At the end of a proposal period, the proposal with most supporters is selected and we move to an exploration period. If there are no proposals, or a tie between proposals, a new proposal period starts.

Exploration period: delegates can cast votes to test or not the winning proposal using the ballot operation. At the end of an exploration period if participation reaches the quorum and the proposal has a supermajority in favor, we proceed to a cooldown period. Otherwise we go back to a proposal period. In any case, if there is enough participation the quorum is updated.

Cooldown period: Nothing happens, this period is only a time gap between exploration and promotion periods. At the end of a cooldown period we move to a promotion period.

Promotion period: delegates can cast votes to promote or not the proposal using the ballot operation. At the end of a promotion period if participation reaches the quorum and the proposal has a supermajority in favor, we move to an adoption period. Otherwise we go back to a proposal period. In any case, if there is enough participation the quorum is updated.

Adoption period: At the end of an adoption period, the proposal is activated as the new protocol.